CaptchaCracker_v2_with_codex/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "captcha-cracker-v2"
version = "0.1.0"
description = "Modernized TensorFlow-based CAPTCHA digit recognizer"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Private User"}
]
dependencies = [
    "tensorflow==2.13.1",
    "tensorflow-io-gcs-filesystem==0.33.0",
    "numpy>=1.24,<3.0",
    "pillow>=10.2",
    "python-dotenv>=1.0",
    "pandas>=2.2",
    "rich>=13.7"
]
